refactor(FT20): change test service structure
